+++ src/react/components/Mutation.tsx
@@ -30,5 +30,5 @@ Mutation.propTypes = {
   children: PropTypes.func.isRequired,
   onCompleted: PropTypes.func,
   onError: PropTypes.func,
-  fetchPolicy: PropTypes.string
-};
+  fetchPolicy: PropTypes.string,
+} as Mutation<any, any>["propTypes"];